image fixed

+++ client/src/app/controllers/containmentCtrl.js
@@ -33,15 +33,20 @@ app.controller('containmentCtrl', ['$scope', '$location', '$http', 'AuthServ', '
         }
 
         $scope.getPipeDataEditPage = function(pipe) {
-            var id = pipe.pipeName ? pipe.pipeName : pipe;
-            $http.get('/getPipeLinebyId/' + id)
-                .success(function(data, status) {
-                    $scope.pipe = data;
-                    // singlePipeData.set($scope.pipe);
-                })
-                .error(function(data, status) {
-                    $scope.pipe = {};
-                });
+            if (pipe.pipeName !== null) {
+                var id = pipe.pipeName ? pipe.pipeName : pipe;
+                $http.get('/getPipeLinebyId/' + id)
+                    .success(function(data, status) {
+                        $scope.pipe = data;
+                        // singlePipeData.set($scope.pipe);
+                    })
+                    .error(function(data, status) {
+                        $scope.pipe = {};
+                    });
+            } else {
+                $scope.pipe = {};
+            }
+
 
         }
 
@@ -74,20 +79,24 @@ app.controller('containmentCtrl', ['$scope', '$location', '$http', 'AuthServ', '
                 });
 
         }
+        
+        $scope.selectedItem = 0;
 
         $scope.deletePipeData = function(removePipeData, index) {
-            $http.delete('/removePipeline/' + removePipeData._id)
-                .success(function(data, status) {
-                    if ($rootScope.user.scope == "Admin") {
-                        $scope.getpipedata.splice(index, 1);
-                        growl.addSuccessMessage('Pipe deleted Successfully');
+            if (removePipeData !== null) {
+                $http.delete('/removePipeline/' + removePipeData._id)
+                    .success(function(data, status) {
+                        if ($rootScope.user.scope == "Admin") {
+                            $scope.getpipedata.splice(index, 1);
+                            growl.addSuccessMessage('Pipe deleted Successfully');
 
-                    }
-                })
-                .error(function(data, status) {
-                    growl.addErrorMessage(data.message);
-                });
+                        }
+                    })
+                    .error(function(data, status) {
+                        growl.addErrorMessage(data.message);
+                    });
 
+            }
         }
 
         //Rupture shape and area 
@@ -96,31 +105,34 @@ app.controller('containmentCtrl', ['$scope', '$location', '$http', 'AuthServ', '
 
         $scope.result = false;
         $scope.calcVolume = function(area, shape) {
+            inchtoMeter = 0.0254;
             $scope.result = true;
             if (shape == "Rectangular") {
-                var inchtoMeter = 0.0254;
                 res = area.length * inchtoMeter * inchtoMeter * area.breadth;
-                console.log(res);
-                $scope.spillVol = spillVolume(res);
-                $scope.spillLitres = $scope.spillVol/0.0010000;
-
-
+                spillVolume(res);
 
             } else
             if (shape == "Triangular") {
                 s = (area.sidea + area.sideb + area.sidec) / 2;
-                result = Math.sqrt(s * (s - area.sidea) * (s - area.sideb) * (s - area.sidec));
-                $scope.getArea = $scope.getArea = parseFloat(Math.round(result * 100) / 100).toFixed(2);
+                res = Math.sqrt(s * (s - area.sidea) * (s - area.sideb) * (s - area.sidec));
+                spillVolume(res);
+
             } else
-            if (shape == "Square")
-                $scope.getArea = area.length * area.length;
-            else
+            if (shape == "Square") {
+                res = area.length * inchtoMeter * inchtoMeter * area.length;
+                spillVolume(res);
+            } else
             if (shape == "Circular") {
-                result = Math.PI * area.radius * area.radius;
-                $scope.getArea = parseFloat(Math.round(result * 100) / 100).toFixed(2);
+                res = Math.PI * area.radius * inchtoMeter * inchtoMeter * area.radius;
+                spillVolume(res);
+
             } else
-            if (shape == "Sector")
-                $scope.getArea = (area.radius * area.angle * Math.PI) / 360;
+            if (shape == "Sector") {
+                res = (area.radius * inchtoMeter * area.angle * Math.PI) / 360;
+                spillVolume(res);
+            }
+
+
         }
 
         var spillVolume = function(area) {
@@ -128,9 +140,9 @@ app.controller('containmentCtrl', ['$scope', '$location', '$http', 'AuthServ', '
                 velocity, pressure = 20;
             var GRAVITY = 9.8; //In meter/second square
             velocity = Math.sqrt(2 * GRAVITY * pipeHeight);
-            result = area * velocity;
-            console.log(result);
-            return result;
+            $scope.spillVol = parseFloat(Math.round(area * velocity * 0.00838641436 * 3600 * 100) / 100).toFixed(2);
+            $scope.spillLitres = $scope.spillVol * 1000 * 3600;
+            $scope.barrels = parseFloat(Math.round($scope.spillLitres * 0.00838641436 * 3600 * 100) / 100).toFixed(2);
         }
 
     }
